move ui initialisation from system to monobehaviour
