+++ workers/unity/Assets/Playground/Config/WorkerUtils.cs
@@ -23,7 +23,6 @@ namespace Playground
             world.GetOrCreateSystem<MoveLocalPlayerSystem>();
             world.GetOrCreateSystem<InitCameraSystem>();
             world.GetOrCreateSystem<FollowCameraSystem>();
-            world.GetOrCreateSystem<InitUISystem>();
             world.GetOrCreateSystem<UpdateUISystem>();
             world.GetOrCreateSystem<PlayerCommandsSystem>();
             world.GetOrCreateSystem<MetricSendSystem>();
