+++ workers/unity/Assets/Playground/Resources/Prefabs/Common/Character.prefab
@@ -17,6 +17,7 @@ GameObject:
   - component: {fileID: 114602415406350378}
   - component: {fileID: 2643701918418317171}
   - component: {fileID: -663445870753195477}
+  - component: {fileID: 8124006586705730658}
   m_Layer: 9
   m_Name: Character
   m_TagString: Untagged
@@ -147,6 +148,18 @@ MonoBehaviour:
   m_Enabled: 0
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 5053c3754e6a48118a61672cedc67dec, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &8124006586705730658
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1292271339760838}
+  m_Enabled: 0
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 682c3f80479014542ab23d20341b0d2e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &1494660013534148
