+++ workers/unity/Assets/Playground/Scripts/UI/InitUIBehaviour.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 88b259c5d0d033d44a239d62a5e91fd8
+guid: 682c3f80479014542ab23d20341b0d2e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
